requirements.txt
fastapi
uvicorn
numpy
python-multipart
opencv-python
keras.models
tensorflow
Pillow
firebase-admin